+++ fuzz/fuzz_targets/flacfile_read_from.rs
@@ -3,8 +3,11 @@
 use std::io::Cursor;
 
 use libfuzzer_sys::fuzz_target;
-use lofty::AudioFile;
+use lofty::{AudioFile, ParseOptions};
 
 fuzz_target!(|data: Vec<u8>| {
-	let _ = lofty::flac::FlacFile::read_from(&mut Cursor::new(data), false);
+	let _ = lofty::flac::FlacFile::read_from(
+		&mut Cursor::new(data),
+		ParseOptions::new().read_properties(false),
+	);
 });