+++ src/macros.rs
@@ -68,4 +68,45 @@ macro_rules! decode_err {
 	};
 }
 
-pub(crate) use {decode_err, err, try_vec};
+// A macro for handling the different `ParsingMode`s
+//
+// NOTE: All fields are optional, if `STRICT` or `RELAXED` are missing, it will
+// 		 fall through to `DEFAULT`. If `DEFAULT` is missing, it will fall through
+// 		 to an empty block.
+//
+// Usage:
+//
+// - parse_mode_choice!(
+// 		ident_of_parsing_mode,
+// 		STRICT: some_expr,
+// 		RELAXED: some_expr,
+// 		DEFAULT: some_expr,
+// 	 )
+macro_rules! parse_mode_choice {
+	(
+		$parse_mode:ident,
+		$(STRICT: $strict_handler:expr,)?
+		$(RELAXED: $relaxed_handler:expr,)?
+		DEFAULT: $default:expr
+	) => {
+		match $parse_mode {
+			$(crate::probe::ParsingMode::Strict => { $strict_handler },)?
+			$(crate::probe::ParsingMode::Relaxed => { $relaxed_handler },)?
+			_ => { $default }
+		}
+	};
+	(
+		$parse_mode:ident,
+		$(STRICT: $strict_handler:expr,)?
+		$(RELAXED: $relaxed_handler:expr $(,)?)?
+	) => {
+		match $parse_mode {
+			$(crate::probe::ParsingMode::Strict => { $strict_handler },)?
+			$(crate::probe::ParsingMode::Relaxed => { $relaxed_handler },)?
+			#[allow(unreachable_patterns)]
+			_ => { unreachable!() }
+		}
+	};
+}
+
+pub(crate) use {decode_err, err, parse_mode_choice, try_vec};