+++ src/ogg/speex/mod.rs
@@ -5,6 +5,7 @@ use super::tag::VorbisComments;
 use crate::error::Result;
 use crate::file::AudioFile;
 use crate::ogg::constants::SPEEXHEADER;
+use crate::probe::ParseOptions;
 use crate::tag::TagType;
 use properties::SpeexProperties;
 
@@ -30,14 +31,14 @@ pub struct SpeexFile {
 impl AudioFile for SpeexFile {
 	type Properties = SpeexProperties;
 
-	fn read_from<R>(reader: &mut R, read_properties: bool) -> Result<Self>
+	fn read_from<R>(reader: &mut R, parse_options: ParseOptions) -> Result<Self>
 	where
 		R: Read + Seek,
 	{
 		let file_information = super::read::read_from(reader, SPEEXHEADER, &[])?;
 
 		Ok(Self {
-            properties: if read_properties { properties::read_properties(reader, &file_information.1)? } else { SpeexProperties::default() },
+            properties: if parse_options.read_properties { properties::read_properties(reader, &file_information.1)? } else { SpeexProperties::default() },
             #[cfg(feature = "vorbis_comments")]
             // Safe to unwrap, a metadata packet is mandatory in Speex
             vorbis_comments_tag: file_information.0.unwrap(),