+++ src/properties.rs
@@ -86,6 +86,7 @@ mod tests {
 	use crate::ogg::{
 		OpusFile, OpusProperties, SpeexFile, SpeexProperties, VorbisFile, VorbisProperties,
 	};
+	use crate::probe::ParseOptions;
 	use crate::wavpack::{WavPackFile, WavPackProperties};
 	use crate::{AudioFile, FileProperties};
 
@@ -274,7 +275,7 @@ mod tests {
 	{
 		let mut f = File::open(path).unwrap();
 
-		let audio_file = T::read_from(&mut f, true).unwrap();
+		let audio_file = T::read_from(&mut f, ParseOptions::default()).unwrap();
 
 		audio_file.properties().clone()
 	}