+++ tests/files/aiff.rs
@@ -1,11 +1,15 @@
 use crate::{set_artist, temp_file, verify_artist};
-use lofty::{FileType, ItemKey, ItemValue, TagExt, TagItem, TagType};
+use lofty::{FileType, ItemKey, ItemValue, ParseOptions, Probe, TagExt, TagItem, TagType};
 use std::io::{Seek, Write};
 
 #[test]
 fn read() {
 	// Here we have an AIFF file with both an ID3v2 chunk and text chunks
-	let file = lofty::read_from_path("tests/files/assets/minimal/full_test.aiff", false).unwrap();
+	let file = Probe::open("tests/files/assets/minimal/full_test.aiff")
+		.unwrap()
+		.options(ParseOptions::new().read_properties(false))
+		.read()
+		.unwrap();
 
 	assert_eq!(file.file_type(), FileType::AIFF);
 
@@ -20,7 +24,12 @@ fn read() {
 fn write() {
 	let mut file = temp_file!("tests/files/assets/minimal/full_test.aiff");
 
-	let mut tagged_file = lofty::read_from(&mut file, false).unwrap();
+	let mut tagged_file = Probe::new(&mut file)
+		.options(ParseOptions::new().read_properties(false))
+		.guess_file_type()
+		.unwrap()
+		.read()
+		.unwrap();
 
 	assert_eq!(tagged_file.file_type(), FileType::AIFF);
 
@@ -32,7 +41,12 @@ fn write() {
 
 	// Now reread the file
 	file.rewind().unwrap();
-	let mut tagged_file = lofty::read_from(&mut file, false).unwrap();
+	let mut tagged_file = Probe::new(&mut file)
+		.options(ParseOptions::new().read_properties(false))
+		.guess_file_type()
+		.unwrap()
+		.read()
+		.unwrap();
 
 	crate::set_artist!(tagged_file, primary_tag_mut, "Bar artist", 1 => file, "Foo artist");