+++ tests/files/mpeg.rs
@@ -1,11 +1,17 @@
 use crate::{set_artist, temp_file, verify_artist};
-use lofty::{Accessor, FileType, ItemKey, ItemValue, TagExt, TagItem, TagType};
+use lofty::{
+	Accessor, FileType, ItemKey, ItemValue, ParseOptions, Probe, TagExt, TagItem, TagType,
+};
 use std::io::{Seek, Write};
 
 #[test]
 fn read() {
 	// Here we have an MP3 file with an ID3v2, ID3v1, and an APEv2 tag
-	let file = lofty::read_from_path("tests/files/assets/minimal/full_test.mp3", false).unwrap();
+	let file = Probe::open("tests/files/assets/minimal/full_test.mp3")
+		.unwrap()
+		.options(ParseOptions::new().read_properties(false))
+		.read()
+		.unwrap();
 
 	assert_eq!(file.file_type(), FileType::MPEG);
 
@@ -22,8 +28,10 @@ fn read() {
 #[test]
 fn read_with_junk_bytes_between_frames() {
 	// Read a file that includes an ID3v2.3 data block followed by four bytes of junk data (0x20)
-	let file =
-		lofty::read_from_path("tests/files/assets/junk_between_id3_and_mp3.mp3", true).unwrap();
+	let file = Probe::open("tests/files/assets/junk_between_id3_and_mp3.mp3")
+		.unwrap()
+		.read()
+		.unwrap();
 
 	// note that the file contains ID3v2 and ID3v1 data
 	assert_eq!(file.file_type(), FileType::MPEG);
@@ -47,7 +55,12 @@ fn read_with_junk_bytes_between_frames() {
 fn write() {
 	let mut file = temp_file!("tests/files/assets/minimal/full_test.mp3");
 
-	let mut tagged_file = lofty::read_from(&mut file, false).unwrap();
+	let mut tagged_file = Probe::new(&mut file)
+		.options(ParseOptions::new().read_properties(false))
+		.guess_file_type()
+		.unwrap()
+		.read()
+		.unwrap();
 
 	assert_eq!(tagged_file.file_type(), FileType::MPEG);
 
@@ -62,7 +75,12 @@ fn write() {
 
 	// Now reread the file
 	file.rewind().unwrap();
-	let mut tagged_file = lofty::read_from(&mut file, false).unwrap();
+	let mut tagged_file = Probe::new(&mut file)
+		.options(ParseOptions::new().read_properties(false))
+		.guess_file_type()
+		.unwrap()
+		.read()
+		.unwrap();
 
 	crate::set_artist!(tagged_file, primary_tag_mut, "Bar artist", 1 => file, "Foo artist");