+++ CHANGELOG.md
@@ -7,10 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `ParseOptions`:
+  - ⚠️ Important ⚠️: This update introduces `ParseOptions` to allow for finer grained control over error
+      eagerness and other settings. Previously, when reading a file the only option available was
+      `read_properties`, specified with a `bool` in `read_from{_path}`. This will now default to `true`,
+      and can be overridden when using `Probe`.
 - **FileProperties**: `FileProperties::new`
 
 ### Changed
 - **ID3v2**: Frame/tag flags with optional additional data are now `Option<T>` instead of `(bool, T)`
+- `read_from{_path}` will no longer take a `bool` for reading properties, and will do it by default. To
+  change this behavior, you must now use `Probe`.
 
 ## [0.8.1] - 2022-09-09